requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
openai==1.58.1
langchain-openai==0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
sqlalchemy==2.0.35
psycopg[binary]==3.2.4
redis==6.2.0
python-dotenv==1.0.1
pydantic==2.9.2
aiosqlite==0.20.0
apscheduler==3.10.4
python-multipart==0.0.9
slowapi==0.1.9
langchain-huggingface==0.1.2
langchain-postgres==0.0.17
sympy>=1.13
PyGithub==2.5.0
email-validator==2.2.0
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
ruff==0.3.5
